All in 1 shader and a bunch of unfinished stuff

+++ Assets/Scripts/StatusEffects/IModifyDamageStatusEffect.cs
@@ -4,6 +4,6 @@ namespace StatusEffects
 {
     public interface IModifyDamageStatusEffect
     {
-        public void ModifyDamage(Character damager, ref float damage);
+        public void ModifyDamage(Character damager, ref double damage);
     }
 }
